Move source files into "src" folder and output into "lib/index.js"
