+++ rollup.js
@@ -1,6 +1,6 @@
 import babel from 'rollup-plugin-babel';
 
 export default {
-  entry: 'lib/index.js',
+  entry: 'src/index.js',
   plugins: [ babel() ],
 };
