+++ src/concatenator.js
