+++ src/decode.js
