+++ src/decoder.js
