+++ src/encode.js
