+++ src/encoder.js
