+++ src/int-buffer-reader.js
