+++ src/int-buffer-writer.js
